Added 2 Unit Tests
Added two new unit tests. One of them doesn't pass because the functionality it tests isn't currently properly implemented (showing grades in the graph on Assignment Details page). As such I can't say for sure if that test is actually properly implemented itself, so I'll have to check back on that when that functionality does get fixed